Update changelog language to match previous versions
The 'Added' section occupies an exceptional semantic space. Rather
than starting with a present tense verbs as all the other sections do,
it starts with what was added. In general, this is because 'Added'
answers the question 'What was added?' rather than 'How was notes
changed?' or 'How was notes fixed?'